feat: Event 인터페이스 작성
