+++ src/main/java/christmas/domain/event/DdayEvent.java
@@ -2,8 +2,10 @@ package christmas.domain.event;
 
 import christmas.domain.VisitingDate;
 
-public class DdayEvent {
-    public int applyDdayEvent(VisitingDate visitingDate) {
+public class DdayEvent implements Event{
+    @Override
+    public int calculateEventDiscount(EventContext context) {
+        VisitingDate visitingDate = context.getVisitingDate();
         int discount = 0;
         if (!visitingDate.isAfterChristmas()) {
             discount = visitingDate.calculateDiscountPriceForDdayEvent();
