+++ src/main/java/christmas/domain/event/Event.java
@@ -1,0 +1,5 @@
+package christmas.domain.event;
+
+public interface Event {
+    int calculateEventDiscount(EventContext context);
+}
