+++ src/main/java/christmas/domain/event/GiveawayEvent.java
@@ -2,8 +2,10 @@ package christmas.domain.event;
 
 import christmas.domain.Order;
 
-public class GiveawayEvent {
-    public int applyGiveawayEvent(Order order) {
+public class GiveawayEvent implements Event{
+    @Override
+    public int calculateEventDiscount(EventContext context) {
+        Order order = context.getOrder();
         int discount = 0;
         if (order.isPriceForGiveawayEvent()) {
             discount -= 25_000;
